requirements.txt
langchain>=0.1.0
langchain-groq>=0.0.1
faiss-cpu>=1.7.4
pypdf>=3.17.1
huggingface_hub>=0.20.3
sentence-transformers>=2.2.2
fastapi>=0.109.0
uvicorn>=0.27.0
python-dotenv>=1.0.0
streamlit>=1.31.0
requests>=2.31.0 